CRUD POST dengan API

+++ backend/app/Http/Controllers/PostController.php
@@ -5,15 +5,18 @@ namespace App\Http\Controllers;
 use App\Models\Post;
 use App\Http\Requests\StorePostRequest;
 use App\Http\Requests\UpdatePostRequest;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\Gate;
 
-class PostController extends Controller
+class PostController extends Controller implements HasMiddleware
 {
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
-        return Post::all();
+        return Post::with('user')->latest()->get();
     }
 
     /**
@@ -23,8 +26,9 @@ class PostController extends Controller
     {
         $validate = $request->validated();
 
-        $post = Post::create($validate);
-        return $post;
+        // $post = Post::create($validate);
+        $post = $request->user()->posts()->create($validate);
+        return ['post' => $post, 'user' => $post->user];
     }
 
     /**
@@ -32,7 +36,7 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
-        return $post;
+        return ['post' => $post, 'user' => $post->user];
     }
 
     /**
@@ -40,10 +44,12 @@ class PostController extends Controller
      */
     public function update(UpdatePostRequest $request, Post $post)
     {
+        Gate::authorize('modify', $post);
+
         $validate = $request->validated();
 
         $post->update($validate);
-        return $post;
+        return ['post' => $post, 'user' => $post->user];
     }
 
     /**
@@ -51,7 +57,15 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        Gate::authorize('modify', $post);
+
         $post->delete();
         return ['message' => 'Post deleted successfully!!'];
+    }
+
+    public static function middleware(){
+        return [
+            new Middleware('auth:sanctum', except: ['index', 'show'])
+        ];
     }
 }
